Resturcture of the vote classes
