+++ app/src/main/java/com/csdfossteam/palermo/core/vote/MajorityVote.java
@@ -1,0 +1,16 @@
+package com.csdfossteam.palermo.core.vote;
+
+/**
+ * TODO doc
+ *
+ * @author Akritas Akritidis
+ */
+public class MajorityVote {
+
+
+
+    public static final class Result {
+
+    }
+
+}
